+++ include/optimproblem.hpp
@@ -20,7 +20,7 @@
 class OptimProblem {
 
   /* ODE stuff */
-  int ninit;                            /* Number of initial conditions to be considered (N^2, N, or 1) */
+  size_t ninit;                          /* Number of initial conditions to be considered (N^2, N, or 1) */
   int ninit_local;                      /* Local number of initial conditions on this processor */
   Vec rho_t0;                            /* Storage for initial condition of the ODE */
   Vec rho_t0_bar;                        /* Adjoint of ODE initial condition */
