+++ src/pythoninterface.cpp
@@ -48,7 +48,7 @@ void PythonInterface::receiveHsys(Mat& Bd){
   MPI_Bcast(vals.data(), nelems, MPI_DOUBLE, 0, MPI_COMM_WORLD);
 
   /* Iterate over all elements*/
-  for (int i = 0; i<vals.size(); i++) {
+  for (size_t i = 0; i<vals.size(); i++) {
     if (fabs(vals[i])<1e-14) continue; // Skip zeros
 
     // Get position in the Bd matrix
@@ -111,7 +111,7 @@ void PythonInterface::receiveHc(int noscillators, std::vector<std::vector<Mat>>&
 
     // Iterate over received elements and place into Bc_vec
     MatGetOwnershipRange(Bc_vec[k][0], &ilow, &iupp);
-    for (int l = 0; l<vals.size(); l++) {
+    for (size_t l = 0; l<vals.size(); l++) {
       if (fabs(vals[l])<1e-14) continue; // Skip zeros
 
       // Get position in the Bc matrix
@@ -153,7 +153,7 @@ void PythonInterface::receiveHc(int noscillators, std::vector<std::vector<Mat>>&
 
     // Iterate over received vals and place into Ac_vec
     MatGetOwnershipRange(Ac_vec[k][0], &ilow, &iupp);
-    for (int l = 0; l<vals.size(); l++) {
+    for (size_t l = 0; l<vals.size(); l++) {
       if (fabs(vals[l])<1e-14) continue; // Skip zeros
 
       // Get position in the Ac matrix
